// src/interface/weeklyreport/type.ts
export interface ICreateInvalidTimesheetArgs {
  employee_code: number;
  invalid_date: string;
  invalid_type: string;
  shift_from: string;
  shift_to: string;
  shift_break: string;
  real_time_attendance_data: string;
  validation_data: string;
  reason: string;
  remarks: string;
  owner: number;
  reviewer: number;
  validated: string;
  keep_in_time: boolean;
  attendance_missing_from: string;
  attendance_missing_to: string;
}
export interface IInvalidTimesheet {
  id: number;
  employee_id: any;
  employee_code: string;
  employee_name: string;
  department: string;
  position: string;
  invalid_type: string;
  reason: string;
  reviewer: any;
  invalid_date: string;
  real_time_attendance_data: string;
  validation_data: string;
  shift_from: string;
  shift_to: string;
  shift_break: string;
  validated: string;
  remarks: string;
  page_size: number;
  page: number;
  create_date: string;
  keep_in_time: boolean;
  date_from: string;
  date_to: string;
  company_id: number;
  attendance_missing_from: string;
  attendance_missing_to: string;
  [key:string]:any;
}